move fdSite class to its corresponding cpp and h files
