+++ fract4d/c/fract4dmodule.cpp
@@ -32,6 +32,7 @@
 #include "fract4dc/colormaps.h"
 #include "fract4dc/loaders.h"
 #include "fract4dc/pysite.h"
+#include "fract4dc/fdsite.h"
 
 struct module_state
 {
@@ -210,17 +211,6 @@ cmap_pylookup_with_flags(PyObject *self, PyObject *args)
 * Sites
 */
 
-typedef enum
-{
-    ITERS,
-    IMAGE,
-    PROGRESS,
-    STATUS,
-    PIXEL,
-    TOLERANCE,
-    STATS,
-} msg_type_t;
-
 static IImage *
 image_fromcapsule(PyObject *pyimage)
 {
@@ -321,147 +311,6 @@ struct calc_args
     }
 };
 
-// write the callbacks to a file descriptor
-class FDSite : public IFractalSite
-{
-public:
-    FDSite(int fd_) : fd(fd_), tid((pthread_t)0),
-                      interrupted(false), params(NULL)
-    {
-#ifdef DEBUG_CREATION
-        fprintf(stderr, "%p : FD : CTOR\n", this);
-#endif
-        pthread_mutex_init(&write_lock, NULL);
-    }
-
-    inline void send(msg_type_t type, int size, void *buf)
-    {
-        pthread_mutex_lock(&write_lock);
-        if (write(fd, &type, sizeof(type)))
-        {
-        };
-        if (write(fd, &size, sizeof(size)))
-        {
-        };
-        if (write(fd, buf, size))
-        {
-        };
-        pthread_mutex_unlock(&write_lock);
-    }
-    virtual void iters_changed(int numiters)
-    {
-        send(ITERS, sizeof(int), &numiters);
-    }
-    virtual void tolerance_changed(double tolerance)
-    {
-        send(TOLERANCE, sizeof(tolerance), &tolerance);
-    }
-
-    // we've drawn a rectangle of image
-    virtual void image_changed(int x1, int y1, int x2, int y2)
-    {
-        if (!interrupted)
-        {
-            int buf[4] = {x1, y1, x2, y2};
-            send(IMAGE, sizeof(buf), &buf[0]);
-        }
-    }
-    // estimate of how far through current pass we are
-    virtual void progress_changed(float progress)
-    {
-        if (!interrupted)
-        {
-            int percentdone = (int)(100.0 * progress);
-            send(PROGRESS, sizeof(percentdone), &percentdone);
-        }
-    }
-
-    virtual void stats_changed(pixel_stat_t &stats)
-    {
-        if (!interrupted)
-        {
-            send(STATS, sizeof(stats), &stats);
-        }
-    }
-
-    // one of the status values above
-    virtual void status_changed(int status_val)
-    {
-        send(STATUS, sizeof(status_val), &status_val);
-    }
-
-    // return true if we've been interrupted and are supposed to stop
-    virtual bool is_interrupted()
-    {
-        //fprintf(stderr,"int: %d\n",interrupted);
-        return interrupted;
-    }
-
-    // pixel changed
-    virtual void pixel_changed(
-        const double *params, int maxIters, int nNoPeriodIters,
-        int x, int y, int aa,
-        double dist, int fate, int nIters,
-        int r, int g, int b, int a)
-    {
-        /*
-	    fprintf(stderr,"pixel: <%g,%g,%g,%g>(%d,%d,%d) = (%g,%d,%d)\n",
-		   params[0],params[1],params[2],params[3],
-		   x,y,aa,dist,fate,nIters);
-	    */
-        return; // FIXME
-    };
-
-    virtual void interrupt()
-    {
-#ifdef DEBUG_THREADS
-        fprintf(stderr, "%p : CA : INT(%p)\n", this, tid);
-#endif
-        interrupted = true;
-    }
-
-    virtual void start(calc_args *params_)
-    {
-#ifdef DEBUG_THREADS
-        fprintf(stderr, "clear interruption\n");
-#endif
-        interrupted = false;
-        params = params_;
-    }
-
-    virtual void set_tid(pthread_t tid_)
-    {
-#ifdef DEBUG_THREADS
-        fprintf(stderr, "%p : CA : SET(%p)\n", this, tid_);
-#endif
-        tid = tid_;
-    }
-
-    virtual void wait()
-    {
-        if (tid != 0)
-        {
-#ifdef DEBUG_THREADS
-            fprintf(stderr, "%p : CA : WAIT(%p)\n", this, tid);
-#endif
-            pthread_join(tid, NULL);
-        }
-    }
-    ~FDSite()
-    {
-#ifdef DEBUG_CREATION
-        fprintf(stderr, "%p : FD : DTOR\n", this);
-#endif
-        close(fd);
-    }
-
-private:
-    int fd;
-    pthread_t tid;
-    volatile bool interrupted;
-    calc_args *params;
-    pthread_mutex_t write_lock;
-};
 
 static void
 site_delete(IFractalSite *site)
